requirements.txt
opencv-python
requests
Pillow
python-dotenv
torch
torchvision
